spark-stream/producer/requirements.txt
websocket-client
kafka-python
web3